Fix Vol info update + add Drive and InChanger to Media rec
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@807 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/baconfig.h
@@ -434,7 +434,7 @@ extern int thr_setconcurrency(int);
 
 #ifdef HAVE_DARWIN_OS
 /* Apparently someone forgot to wrap getdomainname as a C function */
-extern "C" int getdomainname(char *name, size_t len);
+extern "C" int getdomainname(char *name, int len);
 
 /* Darwin lib fnmatch() doesn't work, so use our own */
 #undef HAVE_FNMATCH
